Allow multiple runs to reduce p-value variance.
Reduced p-value variance improves quality of produced labels.

Adds n_iters parameter to control number of runs. Currently defaults to 1 (in which case behavior is unchanged), but recommended to set n_iters=5. NOTE: This will likely be the default value in a future release.
Also note that many public attributes are unset when n_iters > 1. See docstring.

Adds p_values_ attribute, providing the confidence that each cell is a doublet.